requirements.txt
webdriver-manager
selenium
python-dotenv
lxml>=5.1.0
